refactor: extract settings, brushed, and runtime_loop modules
Further behavior-neutral split of main.c:

- settings.c: load/save EEPROM apply + bootloader checkDeviceInfo
- brushed.c: BRUSHED_MODE control loop
- runtime_loop.c: variable PWM, desync, IRQ priority, telemetry,
  ADC/LVC, and sine/stepper vs brushless running branch

main.c drops to ~840 lines (init + thin main loop).
