+++ Inc/brushed.h
@@ -1,0 +1,11 @@
+/*
+ * brushed.h - two-channel brushed control loop
+ */
+#ifndef BRUSHED_H_
+#define BRUSHED_H_
+
+#ifdef BRUSHED_MODE
+void runBrushedLoop(void);
+#endif
+
+#endif /* BRUSHED_H_ */
