+++ Src/main.c
@@ -232,8 +232,11 @@ an settings option)
 #include "bemf_zc.h"
 #include "control_loop.h"
 #include "faults.h"
+#include "settings.h"
+#include "brushed.h"
+#include "runtime_loop.h"
 
-/* Control path: commutation.c, bemf_zc.c, control_loop.c (split from main). */
+/* Control path: commutation, bemf_zc, control_loop, faults, settings, runtime_loop. */
 #include "signal.h"
 #include "sounds.h"
 #include "targets.h"
@@ -257,7 +260,6 @@ an settings option)
 
 #include "version.h"
 
-void zcfoundroutine(void);
 
 // firmware build options !! fixed speed and duty cycle modes are not to be used
 // with sinusoidal startup !!
@@ -479,20 +481,7 @@ uint16_t readings[50];
 
 uint8_t bemf_timeout_happened = 0;
 uint8_t changeover_step = 5;
-// Zero-cross confirmation counts. On the F051 the confirm loop samples the
-// comparator ~3.5x faster with getCompOutputLevel inlined into the
-// RAM-resident ISR (~16 vs ~56 cycles per sample, the out-of-line call went
-// through a RAM-to-flash veneer), so the counts are scaled up to keep the
-// same wall-clock back EMF sampling window as before the inlining.
-#ifdef MCU_F051
-#define ZC_FILTER_MAX 42
-#define ZC_FILTER_RUN_MIN 10
-#define ZC_FILTER_FAST 7
-#else
-#define ZC_FILTER_MAX 12
-#define ZC_FILTER_RUN_MIN 3
-#define ZC_FILTER_FAST 2
-#endif
+/* ZC filter level thresholds live in runtime_loop.c */
 uint8_t filter_level = 5;
 volatile uint8_t running = 0;
 uint16_t advance = 0;
@@ -602,360 +591,8 @@ volatile char input_ready = 0;
 #endif
 
 
-void loadEEpromSettings()
-{
-    read_flash_bin(eepromBuffer.buffer, eeprom_address, sizeof(eepromBuffer.buffer));
-    if(eepromBuffer.eeprom_version < EEPROM_VERSION){
-      eepromBuffer.max_ramp = 160;    // 0.1% per ms to 25% per ms 
-      eepromBuffer.minimum_duty_cycle = 1; // 0.2% to 51 percent
-      eepromBuffer.disable_stick_calibration = 0; // 
-      eepromBuffer.absolute_voltage_cutoff = 10;  // voltage level 1 to 100 in 0.5v increments
-      eepromBuffer.current_P = 100; // 0-255
-      eepromBuffer.current_I = 0; // 0-255
-      eepromBuffer.current_D = 100; // 0-255
-      eepromBuffer.active_brake_power = 0; // 1-5 percent duty cycle
-      eepromBuffer.reserved_eeprom_3[0] = 0; //14-16  for crsf input
-      eepromBuffer.reserved_eeprom_3[1] = 0;
-      eepromBuffer.reserved_eeprom_3[2] = 0;
-      eepromBuffer.reserved_eeprom_3[3] = 0;
-    }
-    // eepromBuffer.advance_level can either be set to 0-3 with config tools less than 1.90 or 10-42 with 1.90 or above 
-    if (eepromBuffer.advance_level > 42 || (eepromBuffer.advance_level < 10 && eepromBuffer.advance_level > 3)){
-        temp_advance = 16;
-    }
-    if (eepromBuffer.advance_level < 4) {         // old format needs to be converted to 0-32 range
-        temp_advance = (eepromBuffer.advance_level<<3);
-        eepromBuffer.advance_level = temp_advance + 10;
-    }
-    if (eepromBuffer.advance_level < 43 && eepromBuffer.advance_level > 9 ) { // new format subtract 10 from advance
-        temp_advance = eepromBuffer.advance_level - 10;
-    }
 
-    if (eepromBuffer.pwm_frequency < 145 && eepromBuffer.pwm_frequency > 7) {
-      int divider = eepromBuffer.pwm_frequency * 100 / 6;
-      TIMER1_MAX_ARR =   TIM1_AUTORELOAD * 400 / divider;
-      SET_AUTO_RELOAD_PWM(TIMER1_MAX_ARR);
-    } else {
-      tim1_arr = TIM1_AUTORELOAD;
-      SET_AUTO_RELOAD_PWM(tim1_arr);
-    }
-    if(eepromBuffer.minimum_duty_cycle < 51 && eepromBuffer.minimum_duty_cycle > 0){
-    minimum_duty_cycle = eepromBuffer.minimum_duty_cycle * 10;
-    }else{
-    minimum_duty_cycle = 0;
-    }
-    if (eepromBuffer.startup_power < 151 && eepromBuffer.startup_power > 49) {
-            min_startup_duty = minimum_duty_cycle + eepromBuffer.startup_power;
-    } else {
-        min_startup_duty = minimum_duty_cycle;
-    }
-    startup_max_duty_cycle = minimum_duty_cycle + 400;  
-
-    motor_kv = (eepromBuffer.motor_kv * 40) + 20;
-#ifdef THREE_CELL_MAX
-		motor_kv =  motor_kv / 2;
-#endif
-#ifdef ONE_TWO_CELL_MAX
-		motor_kv =  motor_kv / 16;
-#endif
-    setVolume(2);
-    if (eepromBuffer.eeprom_version > 0) { // these commands weren't introduced until eeprom version 1.
-#ifdef CUSTOM_RAMP
-
-#else
-        if (eepromBuffer.beep_volume > 11) {
-            setVolume(5);
-        } else {
-            setVolume(eepromBuffer.beep_volume);
-        }
-#endif
-        servo_low_threshold = (eepromBuffer.servo.low_threshold * 2) + 750; // anything below this point considered 0
-        servo_high_threshold = (eepromBuffer.servo.high_threshold * 2) + 1750; // anything above this point considered 2000 (max)
-        servo_neutral = (eepromBuffer.servo.neutral) + 1374;
-        servo_dead_band = eepromBuffer.servo.dead_band;
-
-        low_cell_volt_cutoff = eepromBuffer.low_cell_volt_cutoff + 250; // 2.5 to 3.5 volts per cell range
-        
-        
-#ifndef HAS_HALL_SENSORS
-        eepromBuffer.use_hall_sensors = 0;
-#endif
-
-        if (eepromBuffer.sine_mode_changeover_thottle_level < 5 || eepromBuffer.sine_mode_changeover_thottle_level > 25) { // sine mode changeover 5-25 percent throttle
-            eepromBuffer.sine_mode_changeover_thottle_level = 5;
-        }
-        if (eepromBuffer.drag_brake_strength == 0 || eepromBuffer.drag_brake_strength > 10) { // drag brake 1-10
-            eepromBuffer.drag_brake_strength = 10;
-        }
-
-        if (eepromBuffer.driving_brake_strength == 0 || eepromBuffer.driving_brake_strength > 9) { // motor brake 1-9
-            eepromBuffer.driving_brake_strength = 10;
-        }
-
-        if(eepromBuffer.driving_brake_strength < 10){
-            dead_time_override = DEAD_TIME + (150 - (eepromBuffer.driving_brake_strength * 10));
-            if (dead_time_override > 200) {
-                dead_time_override = 200;
-            }
-        min_startup_duty = min_startup_duty + dead_time_override;
-        minimum_duty_cycle = minimum_duty_cycle + dead_time_override;
-        throttle_max_at_low_rpm = throttle_max_at_low_rpm + dead_time_override;
-        startup_max_duty_cycle = startup_max_duty_cycle + dead_time_override;
-#ifdef STMICRO
-        TIM1->BDTR |= dead_time_override;
-#endif
-#ifdef ARTERY
-        TMR1->brk |= dead_time_override;
-#endif
-#ifdef GIGADEVICES
-        TIMER_CCHP(TIMER0) |= dead_time_override;
-#endif
-#ifdef NXP
-    	for (int submodule = 0; submodule <= 2; submodule++) {
-    		FLEXPWM0->SM[submodule].DTCNT0 = PWM_DTCNT0_DTCNT0(dead_time_override);	//PWMA deadtime
-    		FLEXPWM0->SM[submodule].DTCNT1 = PWM_DTCNT1_DTCNT1(dead_time_override);	//PWMB deadtime
-    	}
-#endif
-#ifdef WCH
-            TIM1->BDTR |= dead_time_override;
-#endif
-        }
-        if (eepromBuffer.limits.temperature < 70 || eepromBuffer.limits.temperature > 140) {
-            eepromBuffer.limits.temperature = 255;
-        }
-
-        if (eepromBuffer.limits.current > 0 && eepromBuffer.limits.current <= 100) {
-            use_current_limit = 1;
-        }
-        
-        currentPid.Kp = eepromBuffer.current_P*2;
-        currentPid.Ki = eepromBuffer.current_I;
-        currentPid.Kd = eepromBuffer.current_D*2;
-        
-        if (eepromBuffer.sine_mode_power == 0 || eepromBuffer.sine_mode_power > 10) {
-            eepromBuffer.sine_mode_power = 5;
-        }
-
-        // unsinged int cant be less than 0
-        if (eepromBuffer.input_type < 10) {
-            switch (eepromBuffer.input_type) {
-            case AUTO_IN:
-                dshot = 0;
-                servoPwm = 0;
-                EDT_ARMED = 1;
-                break;
-            case DSHOT_IN:
-                dshot = 1;
-                EDT_ARMED = 1;
-                break;
-            case SERVO_IN:
-                servoPwm = 1;
-                break;
-            case SERIAL_IN:
-                break;
-            case EDTARM_IN:
-                EDT_ARM_ENABLE = 1;
-                EDT_ARMED = 0;
-                dshot = 1;
-                break;
-            };
-        } else {
-            dshot = 0;
-            servoPwm = 0;
-            EDT_ARMED = 1;
-        }
-        
-        if(eepromBuffer.max_ramp < 10){
-          ramp_divider = 9;
-          max_ramp_startup = eepromBuffer.max_ramp;
-          max_ramp_low_rpm = eepromBuffer.max_ramp;
-          max_ramp_high_rpm = eepromBuffer.max_ramp;
-        }else{
-          ramp_divider = 0;
-          if((eepromBuffer.max_ramp / 10) < max_ramp_startup){
-            max_ramp_startup = eepromBuffer.max_ramp / 10;
-          }
-          if((eepromBuffer.max_ramp / 10) < max_ramp_low_rpm){
-            max_ramp_low_rpm = eepromBuffer.max_ramp / 10;
-          }
-          if((eepromBuffer.max_ramp / 10) < max_ramp_high_rpm){
-            max_ramp_high_rpm = eepromBuffer.max_ramp / 10;
-          }
-        }
-        
-        if (motor_kv < 300) {
-            low_rpm_throttle_limit = 0;
-        }
-        // guard divisions for an erased eeprom (motor_poles 0 or 0xff),
-        // ARM hardware division returns 0 but it is UB in C
-        uint8_t rpm_level_div = 0;
-        if (eepromBuffer.motor_poles != 0) {
-            rpm_level_div = 32 / eepromBuffer.motor_poles;
-        }
-        if (rpm_level_div != 0) {
-            low_rpm_level = motor_kv / 100 / rpm_level_div;
-            high_rpm_level = motor_kv / 12 / rpm_level_div;
-        } else {
-            low_rpm_level = 0;
-            high_rpm_level = 0;
-        }
-    }
-    reverse_speed_threshold = map(motor_kv, 300, 3000, 1000, 500);
-    if (eepromBuffer.bi_direction){
-      polling_mode_changeover = POLLING_MODE_THRESHOLD / 2;
-    }else{
-      polling_mode_changeover = POLLING_MODE_THRESHOLD;
-    }
-}
-
-void saveEEpromSettings()
-{
-    save_flash_nolib(eepromBuffer.buffer, sizeof(eepromBuffer.buffer), eeprom_address);
-}
-
-
-
-
-/*
- * @brief 	Called by the COM_TIMER interrupt handler after the set wait time
- * 			This computes how much to advance in a commutation step.
- * 			This disables the COM_TIMER interrupt.
- * 			Then it enables the comparator to generate its interrupt.
- */
-
-/*
- * @brief 	Called by the comparator interrupt handler.
- * 			Disables the comparator interrupt.
- * 			Enables the COM_TIMER and sets it to generate an interrupt after the wait time.
- */
-
-
-
-
-
-
-#ifdef BRUSHED_MODE
-void runBrushedLoop()
-{
-
-    uint16_t brushed_duty_cycle = 0;
-
-    if (brushed_direction_set == 0 && adjusted_input > 48) {
-        if (forward) {
-            allOff();
-            delayMicros(10);
-            twoChannelForward();
-        } else {
-            allOff();
-            delayMicros(10);
-            twoChannelReverse();
-        }
-        brushed_direction_set = 1;
-    }
-
-    brushed_duty_cycle = map(adjusted_input, 48, 2047, 0,
-        (TIMER1_MAX_ARR - (TIMER1_MAX_ARR / 20)));
-
-    if (degrees_celsius > eepromBuffer.limits.temperature) {
-        duty_cycle_maximum = map(degrees_celsius, eepromBuffer.limits.temperature,
-            eepromBuffer.limits.temperature + 20, TIMER1_MAX_ARR / 2, 1);
-    } else {
-        duty_cycle_maximum = TIMER1_MAX_ARR - 50;
-    }
-    if (brushed_duty_cycle > duty_cycle_maximum) {
-        brushed_duty_cycle = duty_cycle_maximum;
-    }
-
-    if (use_current_limit) {
-        use_current_limit_adjust -= (int16_t)(doPidCalculations(&currentPid, actual_current,
-                                                  CURRENT_LIMIT * 100)
-            / 10000);
-        if (use_current_limit_adjust < minimum_duty_cycle) {
-            use_current_limit_adjust = minimum_duty_cycle;
-        }
-
-        if (brushed_duty_cycle > use_current_limit_adjust) {
-            brushed_duty_cycle = use_current_limit_adjust;
-        }
-    }
-    if ((brushed_duty_cycle > 0) && armed) {
-        SET_DUTY_CYCLE_ALL(brushed_duty_cycle);
-        //	  	TIM1->CCR1 = brushed_duty_cycle;
-        //		TIM1->CCR2 = brushed_duty_cycle;
-        //		TIM1->CCR3 = brushed_duty_cycle;
-
-    } else {
-        SET_DUTY_CYCLE_ALL(0);
-        //		TIM1->CCR1 = 0;
-        //// 		TIM1->CCR2 = 0; 		TIM1->CCR3 = 0;
-        brushed_direction_set = 0;
-    }
-}
-#endif
-
-
-/*
-  check device info from the bootloader, confirming pin code and eeprom location
- */
-static void __attribute__((noinline)) checkDeviceInfo(void)
-{
-#ifdef MCU_SITL
-    // no bootloader device info page in SITL
-    return;
-#endif
-#ifdef NXP
-    uint32_t pflashBlockBase  = 0U;
-    uint32_t pflashTotalSize  = 0U;
-    uint32_t pflashSectorSize = 0U;
-
-    //Get flash properties
-    FLASH_API->flash_get_property(&s_flashDriver, kFLASH_PropertyPflashBlockBaseAddr, &pflashBlockBase);
-    FLASH_API->flash_get_property(&s_flashDriver, kFLASH_PropertyPflashSectorSize, &pflashSectorSize);
-    FLASH_API->flash_get_property(&s_flashDriver, kFLASH_PropertyPflashTotalSize, &pflashTotalSize);
-#else
-#define DEVINFO_MAGIC1 0x5925e3da
-#define DEVINFO_MAGIC2 0x4eb863d9
-
-    // Fixed bootloader address; GCC 12+ -Warray-bounds needs care.
-    struct devinfo {
-        uint32_t magic1;
-        uint32_t magic2;
-        uint8_t deviceInfo[9];
-    };
-    volatile const struct devinfo *devinfo =
-        (volatile const struct devinfo *)(uintptr_t)(0x1000u - 32u);
-#if defined(__GNUC__) && (__GNUC__ >= 12)
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Warray-bounds"
-#pragma GCC diagnostic ignored "-Wstringop-overread"
-#endif
-    const uint32_t magic1 = devinfo->magic1;
-    const uint32_t magic2 = devinfo->magic2;
-    const uint8_t eeprom_code = devinfo->deviceInfo[4];
-#if defined(__GNUC__) && (__GNUC__ >= 12)
-#pragma GCC diagnostic pop
-#endif
-    if (magic1 != DEVINFO_MAGIC1 || magic2 != DEVINFO_MAGIC2) {
-        // bootloader does not support this feature, nothing to do
-        return;
-    }
-    // change eeprom_address based on the code in the bootloaders device info
-    switch (eeprom_code) {
-        case 0x1f:
-            eeprom_address = 0x08007c00;
-            break;
-        case 0x35:
-            eeprom_address = 0x0800f800;
-            break;
-        case 0x2b:
-            eeprom_address = 0x0801f800;
-            break;
-    }
-#endif
-
-    // TODO: check pin code and reboot to bootloader if incorrect
-
-}
+/* loadEEpromSettings / saveEEpromSettings / checkDeviceInfo -> settings.c */
 
 int main(void)
 {
@@ -1146,33 +783,7 @@ if(zero_crosses < 5){
 
        RELOAD_WATCHDOG_COUNTER();
 
-        uint16_t next_tim1_arr = tim1_arr; // unchanged unless variable_pwm recomputes it below
-        if (eepromBuffer.variable_pwm == 1) {      // uses range defined by pwm frequency setting
-            next_tim1_arr = map(commutation_interval, 96, 200, TIMER1_MAX_ARR / 2,
-                TIMER1_MAX_ARR);
-        }
-        if (eepromBuffer.variable_pwm == 2) {      // uses automatic range
-          if(average_interval < 250 && average_interval > 100){
-            next_tim1_arr = average_interval * (CPU_FREQUENCY_MHZ/9);
-          }
-          if(average_interval < 100 && average_interval > 0){
-            next_tim1_arr = 100 * (CPU_FREQUENCY_MHZ/9);
-         }
-          if((average_interval >= 250) || (average_interval == 0)){
-              next_tim1_arr = 250 * (CPU_FREQUENCY_MHZ/9);
-          }
-        }
-        if (next_tim1_arr != last_tim1_arr) {
-            last_tim1_arr = next_tim1_arr;
-            // recompute the scale at idle priority, outside the mask so no interrupt is
-            // held off across the divide; then publish tim1_arr and the scale together so
-            // the 20khz routine never pairs a new arr with a stale scale
-            const uint32_t next_scale = ((uint32_t)next_tim1_arr << 16) / 2000;
-            __disable_irq();
-            tim1_arr = next_tim1_arr;
-            pwm_to_arr_scale_q16 = next_scale;
-            __enable_irq();
-        }
+        runtimeUpdateVariablePwm(&last_tim1_arr);
         faultPollSignalTimeout();
 #ifdef USE_CUSTOM_LED
         if ((input >= 47) && (input < 1947)) {
@@ -1199,316 +810,11 @@ if(zero_crosses < 5){
         }
 
         faultUpdateBemfTimeoutPolicy();
-        average_interval = e_com_time / 3;
-        if (desync_check && zero_crosses > 10) {
-            if ((getAbsDif(last_average_interval, average_interval) > average_interval >> 1) && (average_interval < 2000)) { // throttle resitricted before zc 20.
-                zero_crosses = 0;
-                desync_happened++;
-                if ((!eepromBuffer.bi_direction && (input > 47)) || commutation_interval > 1000) {
-                    running = 0;
-                }
-                old_routine = 1;
-                if (zero_crosses > 100) {
-                    average_interval = 5000;
-                }
-                last_duty_cycle = min_startup_duty / 2;
-            }
-            desync_check = 0;
-            //	}
-            last_average_interval = average_interval;
-        }
-
-#if !defined(MCU_G031) && !defined(NEED_INPUT_READY)
-#ifdef NXP
-	if (dshot_telemetry && (commutation_interval > DSHOT_PRIORITY_THRESHOLD)) {
-		NVIC_SetPriority(IC_DMA_IRQ_NAME, 0);
-		NVIC_SetPriority(COM_TIMER_IRQ, 1);
-		NVIC_SetPriority(COMP0_IRQ, 1);
-		NVIC_SetPriority(COMP1_IRQ, 1);
-	} else {
-		NVIC_SetPriority(IC_DMA_IRQ_NAME, 1);
-		NVIC_SetPriority(COM_TIMER_IRQ, 0);
-		NVIC_SetPriority(COMP0_IRQ, 0);
-		NVIC_SetPriority(COMP1_IRQ, 0);
-	}
-#else
-        if (dshot_telemetry && (commutation_interval > DSHOT_PRIORITY_THRESHOLD)) {
-             NVIC_SetPriority(IC_DMA_IRQ_NAME, 0);
-             NVIC_SetPriority(COM_TIMER_IRQ, 1);
-             NVIC_SetPriority(COMPARATOR_IRQ, 1);
-         } else {
-             NVIC_SetPriority(IC_DMA_IRQ_NAME, 1);
-             NVIC_SetPriority(COM_TIMER_IRQ, 0);
-             NVIC_SetPriority(COMPARATOR_IRQ, 0);
-         }
-#endif
-#endif
-        if (send_telemetry) {
-#ifdef USE_SERIAL_TELEMETRY
-            makeTelemPackage((int8_t)degrees_celsius, battery_voltage, actual_current,
-                (uint16_t)(consumed_current >> 16), e_rpm);
-            send_telem_DMA(10);
-            send_telemetry = 0;
-#endif
-        } else if(send_esc_info_flag ) {
-           makeInfoPacket();
-           send_telem_DMA(49);
-           send_esc_info_flag = 0;
-        }
-        if (PROCESS_ADC_FLAG == 1) { // for adc and telemetry set adc counter at 1khz loop rate
-#if defined(STMICRO)
-            ADC_DMA_Callback();
-            LL_ADC_REG_StartConversion(ADC1);
-#ifdef USE_ADC_1_2
-          LL_ADC_REG_StartConversion(ADC2);
-#endif          
-            converted_degrees = __LL_ADC_CALC_TEMPERATURE(3300, ADC_raw_temp, LL_ADC_RESOLUTION_12B);
-#endif
-#ifdef MCU_GDE23
-            ADC_DMA_Callback();
-            // converted_degrees = (1.43 - ADC_raw_temp * 3.3 / 4096) * 1000 / 4.3 + 25;
-            converted_degrees = ((int32_t)(357.5581395348837f * (1 << 16)) - ADC_raw_temp * (int32_t)(0.18736373546511628f * (1 << 16))) >> 16;
-            adc_software_trigger_enable(ADC_REGULAR_CHANNEL);
-#endif
-#ifdef ARTERY
-            ADC_DMA_Callback();
-            adc_ordinary_software_trigger_enable(ADC1, TRUE);
-    #ifdef USE_NTC
-            converted_degrees = getNTCDegrees(ADC_raw_ntc);
-    #else     
-            converted_degrees = getConvertedDegrees(ADC_raw_temp);
-    #endif
-#endif
-#ifdef NXP
-            //Call ADC_DMA callback to get raw data
-            ADC_DMA_Callback();
-
-            //Convert temperature data to actual temperature in degrees Celsius
-            converted_degrees = computeTemperature(ADC_raw_temp[0], ADC_raw_temp[1]);
-
-            //Start ADC conversion
-            startADCConversion();
-#endif
-#ifdef WCH
-            startADCConversion( );
-            converted_degrees = getConvertedDegrees(ADC_raw_temp);
-#endif
-            degrees_celsius = converted_degrees;
-#ifdef NXP
-            //MCXA has 16-bit ADC data
-            battery_voltage = ((7 * battery_voltage) + ((ADC_raw_volts * 3300 / 65535 * VOLTAGE_DIVIDER) / 100)) / 8;
-            smoothed_raw_current = getSmoothedCurrent();
-            //Actual current is in 10mA, so 1 = 10mA
-            actual_current = (((smoothed_raw_current * 3300 / 65535) - CURRENT_OFFSET) * 100) / (MILLIVOLT_PER_AMP);
-#else
-            battery_voltage = ((7 * battery_voltage) + ((ADC_raw_volts * 3300 / 4095 * VOLTAGE_DIVIDER) / 100)) >> 3;
-            smoothed_raw_current = getSmoothedCurrent();
-            actual_current = ((smoothed_raw_current * 3300 / 41) - (CURRENT_OFFSET * 100)) / (MILLIVOLT_PER_AMP);
-#endif
-            if (actual_current < 0) {
-                actual_current = 0;
-            }             
-            if (eepromBuffer.low_voltage_cut_off == 1) {  
-                if (battery_voltage < (cell_count * low_cell_volt_cutoff)) {
-                  low_voltage_count++;
-                } else {
-                  if(!LOW_VOLTAGE_CUTOFF){  // if set low cutoff has happened, require power cycle to reset
-                    low_voltage_count = 0;
-                  }
-                }
-            }
-            if (eepromBuffer.low_voltage_cut_off == 2 ){   // absolute cut off
-              if (battery_voltage <  (eepromBuffer.absolute_voltage_cutoff * 50)) {
-                low_voltage_count++;    
-                } else {
-                  if(!LOW_VOLTAGE_CUTOFF){
-                    low_voltage_count = 0;
-                  }
-                }
-            }
-            if (low_voltage_count > (10000 - (stepper_sine * 9900))) {      // 10 second wait before cut-off for low voltage
-              LOW_VOLTAGE_CUTOFF = 1;
-              input = 0;
-              allOff();
-              maskPhaseInterrupts();
-              running = 0;
-              zero_input_count = 0;
-              armed = 0;
-             }
-           
-            PROCESS_ADC_FLAG = 0;
-#ifdef USE_ADC_INPUT
-            if (ADC_raw_input < 10) {
-                zero_input_count++;
-            } else {
-                zero_input_count = 0;
-            }
-#endif
-        }
-#ifdef USE_ADC_INPUT
-        signaltimeout = 0;
-        ADC_smoothed_input = (((10 * ADC_smoothed_input) + ADC_raw_input) / 11);
-        newinput = ADC_smoothed_input / 2;
-        if (newinput > 2000) {
-            newinput = 2000;
-        }
-#endif
-        stuckcounter = 0;
-        if (stepper_sine == 0) {
-
-            e_rpm = running * (600000 / e_com_time); // in tens of rpm
-            k_erpm = e_rpm / 10; // ecom time is time for one electrical revolution in microseconds
-
-            if (low_rpm_throttle_limit) { // some hardware doesn't need this, its on
-                                          // by default to keep hardware / motors
-                                          // protected but can slow down the response
-                                          // in the very low end a little.
-                duty_cycle_maximum = map(k_erpm, low_rpm_level, high_rpm_level, throttle_max_at_low_rpm,
-                    throttle_max_at_high_rpm); // for more performance lower the
-                                               // high_rpm_level, set to a
-                                               // consvervative number in source.
-            }else{
-							duty_cycle_maximum = 2000;
-						}
-
-            if (degrees_celsius > eepromBuffer.limits.temperature) {
-              duty_cycle_maximum = map(degrees_celsius, eepromBuffer.limits.temperature - 10, eepromBuffer.limits.temperature + 10,
-                throttle_max_at_high_rpm / 2, 1);
-            }
-            if (zero_crosses < 100 && commutation_interval > 500) {
-              filter_level = ZC_FILTER_MAX;
-            } else {
-              filter_level = map(average_interval, 100, 500, ZC_FILTER_RUN_MIN, ZC_FILTER_MAX);
-            }
-            if (commutation_interval < 50) {
-              filter_level = ZC_FILTER_FAST;
-            }
-
-            if (eepromBuffer.auto_advance) {
-              auto_advance_level = map(duty_cycle, 100, 2000, 13, 23);
-            }
-
-            /**************** old routine*********************/
-#ifdef CUSTOM_RAMP
-            if (old_routine && running) {
-                maskPhaseInterrupts();
-                getBemfState();
-                if (!zcfound) {
-                    if (rising) {
-                        if (bemfcounter > min_bemf_counts_up) {
-                            zcfound = 1;
-                            zcfoundroutine();
-                        }
-                    } else {
-                        if (bemfcounter > min_bemf_counts_down) {
-                            zcfound = 1;
-                            zcfoundroutine();
-                        }
-                    }
-                }
-            }
-#endif
-            faultHandleBemfIntervalStall();
-        } else { // stepper sine
-
-#ifdef GIMBAL_MODE
-            step_delay = 300;
-            maskPhaseInterrupts();
-            allpwm();
-            if (newinput > 1000) {
-                desired_angle = map(newinput, 1000, 2000, 180, 360);
-            } else {
-                desired_angle = map(newinput, 0, 1000, 0, 180);
-            }
-            if (current_angle > desired_angle) {
-                forward = 1;
-                advanceincrement();
-                delayMicros(step_delay);
-                current_angle--;
-            }
-            if (current_angle < desired_angle) {
-                forward = 0;
-                advanceincrement();
-                delayMicros(step_delay);
-                current_angle++;
-            }
-#else
-
-            if (input > 48 && armed) {
-
-                if (input > 48 && input < 137) { // sine wave stepper
-
-                    if (do_once_sinemode) {
-                        // disable commutation interrupt in case set
-                        DISABLE_COM_TIMER_INT();
-                        maskPhaseInterrupts();
-                        SET_DUTY_CYCLE_ALL(0);
-                        allpwm();
-                        do_once_sinemode = 0;
-                    }
-                    advanceincrement();
-                    step_delay = map(input, 48, 120, 7000 / eepromBuffer.motor_poles, 810 / eepromBuffer.motor_poles);
-                    delayMicros(step_delay);
-                    e_rpm = 600 / step_delay; // in hundreds so 33 e_rpm is 3300 actual erpm
-
-                } else {
-                    do_once_sinemode = 1;
-                    advanceincrement();
-                    if (input > 200) {
-                        phase_A_position = 0;
-                        step_delay = 80;
-                    }
-
-                    delayMicros(step_delay);
-                    if (phase_A_position == 0) {
-                        stepper_sine = 0;
-                        running = 1;
-                        old_routine = 1;
-                        commutation_interval = 9000;
-                        average_interval = 9000;
-                        last_average_interval = average_interval;
-                        SET_INTERVAL_TIMER_COUNT(9000);
-                        zero_crosses = 20;
-                        prop_brake_active = 0;
-                        step = changeover_step;
-                        // comStep(step);// rising bemf on a same as position 0.
-                        if (eepromBuffer.stall_protection) {
-                            last_duty_cycle = stall_protect_minimum_duty;
-                        }
-                        commutate();
-                        generatePwmTimerEvent();
-                    }
-                }
-
-            } else {
-                do_once_sinemode = 1;
-                if (eepromBuffer.brake_on_stop == 1) {
-#ifndef PWM_ENABLE_BRIDGE
-                    prop_brake_duty_cycle =  eepromBuffer.drag_brake_strength * 200;
-                    adjusted_duty_cycle =  tim1_arr - ((prop_brake_duty_cycle * tim1_arr) / 2000);
-                    if(adjusted_duty_cycle < 100){
-                      fullBrake();
-                    }else{
-                      proportionalBrake();
-                      SET_DUTY_CYCLE_ALL(adjusted_duty_cycle);
-                      prop_brake_active = 1;
-                    } 
-#else
-                    // todo add braking for PWM /enable style bridges.
-#endif
-                } else if (eepromBuffer.brake_on_stop == 2){
-                  comStep(2);
-                  SET_DUTY_CYCLE_ALL(DEAD_TIME + ((eepromBuffer.active_brake_power * tim1_arr) / 2000)* 10);
-                }else{
-                   SET_DUTY_CYCLE_ALL(0);
-                   allOff();
-                }
-                e_rpm = 0;
-            }
-
-#endif // gimbal mode
-        } // stepper/sine mode end
-
+        runtimeProcessDesyncCheck();
+        runtimeUpdateDshotIrqPriority();
+        runtimeSendTelemetryIfNeeded();
+        runtimeProcessAdcAndProtections();
+        runtimeMotorModeTick();
 #ifdef BRUSHED_MODE
         runBrushedLoop();
 #endif
